Remove unnecessary sessionsInitialized property
The sessionsInitialized had the same job as sessions.length. However, sessions is a plain JavaScript array and does not trigger QML bindings update (these updates are only triggered by dependencies changes (QML properties)). Because of this, I replaced the plain JS array with a ListModel. This list model causes the QML to update and the session button to update when sessions are loaded.